[feat] Create models `Contact`, `ContactSection`
- Contact는 연락처 1개를 나타내는 객체
- ContactSection은 연락처 목록 TableView에서 Contact의 이름 첫 문자를 기준으로 각각의 section으로 나뉘는 객체
