+++ RxContacts/RxContacts/Model/Contact.swift
@@ -6,16 +6,7 @@
 //
 
 import Foundation
-import RxDataSources
 
-struct ContactSection {
-    var header: String
-    var items: [Int]
-}
-
-extension ContactSection: SectionModelType {
-    init(original: ContactSection, items: [Int]) {
-        self = original
-        self.items = items
-    }
+struct Contact {
+    var fullName: String
 }
